btrfs: add support for new raid1c34 profiles
There are new 3- and 4-copy variants of RAID1, estimated to be merged to
kernel 5.5. Add the two new profiles to the list of recognized ones. As
this builds on the same code as RAID1, only the redundancy level needs
to be adjusted, the rest is done by the existing code.

Signed-off-by: David Sterba <dsterba@suse.com>